pyproject.toml
[project]
name = "alpha-analyst"
version = "1.1.0"
description = "AI-powered financial analysis agent"
authors = [{ name = "Nilesh Auradkar", email = "nilesh.auradkar14@gmail.com" }]
requires-python = ">=3.10,<3.14"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "python-dotenv>=1.1",
    "loguru>=0.7,<0.8",
    "pydantic>=2.11",
    "pydantic-settings>=2.10,<3",
    "langchain>=1.2",
    "langchain-community>=0.4",
    "langchain-text-splitters>=1.1",
    "langgraph>=1.0",
    "crewai>=1.9.1",
    "protobuf>=5.0",
    "tokenizers>=0.20",
    "chromadb>=1.1",
    "yfinance>=1.1",
    "httpx>=0.28",
    "tavily-python>=0.7",
    "torch>=2.2",
    "transformers>=4.0",
    "pytest>=9.0.2",
    "langchain-ollama>=1.0.1",
    "ragas>=0.4.3",
    "tenacity>=9.1.4",
]

[project.scripts]
alpha-analyst = "api.main:run_api"
evaluate = "evaluation:cli"

[dependency-groups]
api = [
    "fastapi>=0.128",
    "uvicorn[standard]>=0.40",
    "prometheus-client>=0.24",
]
dev = [
    "pytest>=9.0.2,<10",
    "pytest-asyncio>=1.3",
    "pytest-cov>=7.0",
    "ruff>=0.14",
    "mypy>=1.9,<2",
]
eval = [
    "deepeval>=3.8",
    "datasets>=4.5",
]


[tool.uv]
default-groups = ["api", "dev", "eval"]

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "agents",
    "api",
    "configs",
    "evaluation",
    "models",
    "observability",
    "rag",
    "tools",
]

# --- Tool Configs (Ruff, Pytest, etc.) ---

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
ignore_missing_imports = true
